@@ -2,20 +2,20 @@ name: socket_io
 description: >
   Port of JS/Node library Socket.io. It enables real-time, bidirectional and 
   event-based communication cross-platform.
-version: 0.9.4
+version: 1.0.0
 author: jumperchen <jumperchen@potix.com>
 homepage: https://www.zkoss.org
 repository: https://github.com/rikulo/socket.io-dart
 issue_tracker: https://github.com/rikulo/socket.io-dart/issues
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  stream: ^2.2.1+1
-  socket_io_common: ' >= 0.9.0+5 < 2.0.0-0'
-  uuid: ^2.0.4
-  logging: ^0.11.3+2
+  stream: ^3.0.0
+  socket_io_common: ^1.0.0-nullsafety.1
+  uuid: ^3.0.4
+  logging: ^1.0.0
 dev_dependencies:
-  test: ">=1.3.0 <2.0.0"
-  pedantic: ^1.9.0
+  test: ^1.16.8
+  pedantic: ^1.11.0
