@@ -13,8 +13,8 @@ import 'package:socket_io/src/engine/connect.dart';
 import 'package:socket_io/src/engine/transport/polling_transport.dart';
 
 class JSONPTransport extends PollingTransport {
-  String head;
-  String foot;
+  late String head;
+  late String foot;
   JSONPTransport(SocketConnect connect) : super(connect) {
     head = '___eio[' +
         (connect.request.uri.queryParameters['j'] ?? '')
@@ -75,7 +75,7 @@ class JSONPTransport extends PollingTransport {
 
     // Go through all the matches and build the result map.
     for (Match match in search.allMatches(query)) {
-      result[decode(match.group(1))] = decode(match.group(2));
+      result[decode(match.group(1)!)] = decode(match.group(2)!);
     }
 
     return result;
