@@ -12,7 +12,7 @@ import 'package:socket_io/src/engine/server.dart';
 import 'package:socket_io/src/util/event_emitter.dart';
 
 class Engine extends EventEmitter {
-  static Engine attach(server, [Map options]) {
+  static Engine attach(server, [Map? options]) {
     var engine = Server(options);
     engine.attachTo(server, options);
     return engine;
